+++ CONTRIBUTING.md
@@ -1,13 +1,9 @@
 ## Contribution guidelines ##
 
 ## Workflow ##
-
-* Develop on a feature branch and send a pull request for review.
 * Assign the pull request to one of the following contacts:
 	* Primary: Joe Hoyle [@joehoyle](https://github.com/joehoyle)
-	* Secondary: 
 
 ## Coding Standards ##
-
 Please follow these recommendations
 [http://codex.wordpress.org/WordPress_Coding_Standards](http://codex.wordpress.org/WordPress_Coding_Standards)
